requirements.txt
asgiref==3.11.0
Django<5.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
sqlparse==0.5.5
psycopg2-binary==2.9.9
